Move metadata, pagination, and routers from utils.common
