+++ EquiTrack/audit/views.py
@@ -13,6 +13,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.renderers import JSONRenderer
 from rest_framework.response import Response
 
+from EquiTrack.views import ExportViewSetDataMixin, FSMTransitionActionMixin
 from audit.exports import AuditorFirmCSVRenderer, EngagementCSVRenderer
 from audit.filters import DisplayStatusFilter, UniqueIDOrderingFilter
 from audit.metadata import AuditBaseMetadata, EngagementMetadata
@@ -30,10 +31,12 @@ from audit.serializers.export import (
     AuditPDFSerializer, MicroAssessmentPDFSerializer, SpecialAuditPDFSerializer, SpotCheckPDFSerializer,)
 from partners.models import PartnerOrganization
 from partners.serializers.partner_organization_v2 import MinimalPartnerOrganizationListSerializer
-from utils.common.views import (
-    ExportViewSetDataMixin, FSMTransitionActionMixin, MultiSerializerViewSetMixin, NestedViewSetMixin,
-    SafeTenantViewSetMixin,)
-from utils.common.pagination import DynamicPageNumberPagination
+from rest_extra.pagination import DynamicPageNumberPagination
+from rest_extra.views import (
+    MultiSerializerViewSetMixin,
+    NestedViewSetMixin,
+    SafeTenantViewSetMixin,
+)
 from vision.adapters.purchase_order import POSynchronizer
